requirements.txt
Flask==2.3.2
Pillow==10.2.0
gunicorn==21.2.0
